multimodal-agents/requirements.txt
openai>=1.3.0
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
jupyter>=1.0.0
matplotlib>=3.7.0
numpy>=1.24.0

